refactor options/config
