+++ pkg/services/grafana-apiserver/builder/common.go
@@ -1,4 +1,4 @@
-package grafanaapiserver
+package builder
 
 import (
 	"net/http"
